Moved motor/rotor functions into rotor.c Removed LED set from the IADC IRQ handler that was for debug converted all iadc doubles to floats, 32-bit precision is plenty added flash() functionality
A rose, for you:

	---,---`---<@

+++ include/iadc.h
@@ -13,4 +13,4 @@
 #define IADC_INPUT_1_BUSALLOC     GPIO_CDBUSALLOC_CDODD0_ADC0
 
 void initIADC(void);
-double iadc_get_result(int i);
+float iadc_get_result(int i);
